affinity skip bugfix, take max of two skill values
In the conditional statement inserting affinities, it skipped the rest of the iteration, also skipping inserting the skill in question. This was a bug.

In the case where Wurm dumps the current skill value and max skill value weirdly, WUSI will now take whichever one is the higher value for the skill.

Added SQL statement to update the playerSkillIds value.